requirements.txt
python-dotenv
imapclient
pytesseract
pdf2image
pdfplumber
supabase
apscheduler

